Rename vdom files and remove the old one
